car-clasificator/pyproject.toml
[tool.poetry]
name = "car-clasificator"
version = "0.1.0"
description = ""
authors = ["KacperKwiat <s24533@pjwstk.edu.pl>, Kasper Serżysko <s24270@pjwstk.edu.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.0.0"
pillow = "^10.4.0"
matplotlib = "^3.9.1"
tqdm = "^4.66.4"
torch = "^2.3.1"
opencv-python = "^4.10.0.84"
torchvision = "^0.18.1"
easyocr = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
